Navigation Assessment: created components structure / finalized assessment.js
